+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `Metadata` class providing basic information for BayBE objects
-- `ParameterMetadata` class providing metadata for `Parameter` objects
+- `MeasurableMetadata` class providing metadata for `Parameter` objects
 
 ## [0.13.2] - 2025-07-09
 ### Changed